refactor: and a couple more
