+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/const/ConstMultiSet.kt
@@ -1,12 +1,12 @@
 package xyz.lbres.kotlinutils.set.multiset.const
 
-import xyz.lbres.kotlinutils.general.tryOrDefault
 import xyz.lbres.kotlinutils.internal.constants.Suppressions
 import xyz.lbres.kotlinutils.set.multiset.MultiSet
 import xyz.lbres.kotlinutils.set.multiset.utils.CountsMap
 import xyz.lbres.kotlinutils.set.multiset.utils.performIntersect
 import xyz.lbres.kotlinutils.set.multiset.utils.performMinus
 import xyz.lbres.kotlinutils.set.multiset.utils.performPlus
+import xyz.lbres.kotlinutils.utils.tryOrDefault
 
 /**
  * [MultiSet] implementation where values of elements are assumed to be constant.
