+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/utils/operators.kt
@@ -1,10 +1,10 @@
 package xyz.lbres.kotlinutils.set.multiset.utils
 
-import xyz.lbres.kotlinutils.general.simpleIf
 import xyz.lbres.kotlinutils.set.multiset.MultiSet
 import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.impl.AbstractMultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.impl.MultiSetImpl
+import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.math.min
 
 /**
