+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/utils/Labelled.kt
@@ -1,0 +1,15 @@
+package xyz.lbres.kotlinutils.utils
+
+import xyz.lbres.kotlinutils.internal.constants.Suppressions
+
+/**
+ * Class to represent a value with an assigned label
+ *
+ * @param value T
+ * @param label [String]
+ */
+@Suppress(Suppressions.UNUSED)
+data class Labelled<T>(
+    val value: T,
+    val label: String
+)
