+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/array/ArrayExtTest.kt
@@ -1,7 +1,7 @@
 package xyz.lbres.kotlinutils.array
 
-import xyz.lbres.kotlinutils.general.simpleIf
 import xyz.lbres.kotlinutils.list.IntList
+import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
