+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/array/IntArrayExtTest.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.kotlinutils.array
 
-import xyz.lbres.kotlinutils.general.simpleIf
+import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.math.absoluteValue
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
